templates/types/llamaindexserver/fastapi/pyproject.toml
[tool]
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Marcus Schiesser <mail@marcusschiesser.de>"]
readme = "README.md"

[tool.poetry.scripts]
generate = "generate:generate_datasource"
dev = "main:run('dev')"
prod = "main:run('prod')"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
python-dotenv = "^1.0.0"
pydantic = "<2.10"
aiostream = "^0.5.2"
llama-index-core = "^0.12.24"
llama-index-embeddings-openai = "^0.3.1"
llama-index-llms-openai = "^0.3.26"
e2b-code-interpreter = "^1.1.1"
xhtml2pdf = "^0.2.17"
markdown = "^3.7"
llama-index-server = "^0.1.5"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pytest = "^8.3.5"
pytest-asyncio = "^0.25.3"

[tool.mypy]
python_version = "3.11"
plugins = "pydantic.mypy"
exclude = ["tests", "venv", ".venv", "output", "config"]
check_untyped_defs = true
warn_unused_ignores = false
show_error_codes = true
namespace_packages = true
ignore_missing_imports = true
follow_imports = "silent"
implicit_optional = true
strict_optional = false
disable_error_code = ["return-value", "assignment"]

[[tool.mypy.overrides]]
module = "app.*"
ignore_missing_imports = false

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
